qcodec2: fix a potential blocking issue
In some cases, the input queue may be blocked when mNumPendingWorks exceed max,
such as downstream element holding the decoder output buffer and not return.
This will lead to pipeline can not change to READY/NULL state. So try to unblock
the wait condition variable when state changing from PAUSED to READY.

Change-Id: If09ea5b1c9dcdcea392cd66eac167c2788cd5c67
Signed-off-by: Shiming Dong <quic_shimdong@quicinc.com>