mlsnpe: Fix the way we storing the userbuffers
The behaviour of the getUserBuffer in SNPE2.0 was changes and now
it returns different pointer address each time. This breaks the logic
were we try to find the old user buffer and replace it with a new one.
With this fix we are storing the buffer in a map for the specific tensor
name and now it doesn't matter the address value.

Change-Id: Iaf986e9d66db72ed0f7c82a9e7bc9da3ffadbc3c