qcodec2vdec: refine the logic of acquiring external buffers
To support "external buffer + interlaced stream" case, both decoder and
C2D will acquire buffers from the same downstream buffer pool. Since the
number of buffers is limited, we need to limit the number of buffers
that decoder can take to avoid there is no buffer when C2D is trying to
acquire.

Change-Id: Iff4481f383a33e579b588eb7be94039d9c2b56de
Signed-off-by: Shiming Dong <quic_shimdong@quicinc.com>