ml: Use postprocessing module caps in transform_caps
Bug:
Caps negotiation was choosing unsupported tensor type
in case where inference plugin supports two tensor types.

Analysis:
transform_caps function was not returing actual module capabilities,
but template capabilities of the pad. The module was initialized
in set_caps and checks if fixated caps are actually supported. But it is
too late and actually supported format was not selected.

This was causing only first supported type by inference plugin to be selected
which may not be the correct supported type by post processing plugins.

Fix:
Move processing module initialization in NULL to READY transition, and
use module caps in transform_caps function.

Change-Id: I6f49a4ffaf43433cc8acdba1a50f01b001638898