gst-timelapse-example:fix buffer leak
- Issue: When pressing Ctrl+C to stop the program, it gets blocked and cannot exit.
- Reason: When the qmmf-sdk exits, it needs to wait for the gstreamer plugin to release all used buffers,otherwise,it will block and check once per second.
In Gstreamer 1.22, when gst_buffer_copy is called, if the memory is shared, the reference count of the source buffer will be +1, it will be released when the copied buffer is freed.
In Gstreamer 1.20, gst_buffer_copy does not increment the reference count of the source buffer, so no such issue occurs.
- Fix: Push the source buffer directly to appsrc without copying it.

Change-Id: If8e1a4df4cbdf86f19830876b32865a2d2f7a089
Signed-off-by: lei shen <leishen@qti.qualcomm.com>